+++ .vscode/launch.json
@@ -10,7 +10,7 @@
 			"internalConsoleOptions": "neverOpen",
 			"request": "launch",
 			"name": "Debug",
-			"program": "index.ts",
+			"program": "src/index.ts",
 			"cwd": "${workspaceFolder}",
 			"stopOnEntry": false,
 			"watchMode": true
@@ -20,7 +20,7 @@
 			"internalConsoleOptions": "neverOpen",
 			"request": "launch",
 			"name": "Run",
-			"program": "index.ts",
+			"program": "src/index.ts",
 			"cwd": "${workspaceFolder}",
 			"noDebug": true,
 			"watchMode": false
