+++ src/views/reminders.ejs
@@ -42,6 +42,15 @@
                     </div>
                 <% }); %>
             <% } %>
+            <div class="instructions">
+                <p>You can add a reminder via Mira AI by saying<br>
+                    <b>"Hey Mira, remind me to ..."</b><br>
+                    Some other things you can say:<br>
+                    <b>"Hey Mira, mark my reminder to ... as complete"</b><br>
+                    <b>"Hey Mira, delete my reminders about ..."</b><br>
+                    <b>"Hey Mira, what are my reminders?"</b>
+                </p>
+            </div>
         </div>
     </div>
